requirements.txt
langchain
unstructured
openai
llama-index==0.10.27
langchain==0.1.13
llama-index-embeddings-azure-openai 
llama-index-llms-azure-openai
llama-index-vector-stores-postgres
llama-index-vector-stores-chroma
chromadb
llama_index[langchain]
uvicorn
fastapi
rank-bm25
nltk
llama-index-postprocessor-cohere-rerank
scikit-learn
pytest 
httpx
llama-parsei
cohere
